requirements.txt
PyYAML~=5.4.1
numpy~=1.21.0
casadi~=3.5.5
dacite~=1.6.0
matplotlib~=3.4.2